Partial revert of 111fc6a34375ad0d964891bb925d132770739c8d
This seems to work around some compiler bug (or UB consequences? the calculation here shouldn't trigger UB regardless) which causes graphical issues.

+++ src/GPU3D_Soft.h
@@ -123,13 +123,13 @@ private:
             this->x = x;
             if (xdiff != 0 && !linear)
             {
-                u32 num = (x * w0n) << shift;
-                u32 den = (x * w0d) + ((xdiff-x) * w1d);
+                s64 num = ((s64)x * w0n) << shift;
+                s32 den = (x * w0d) + ((xdiff-x) * w1d);
 
                 // this seems to be a proper division on hardware :/
                 // I haven't been able to find cases that produce imperfect output
                 if (den == 0) yfactor = 0;
-                else          yfactor = num / den;
+                else          yfactor = (s32)(num / den);
             }
         }
